Restructure format A
