+++ source/io/formatA/commands.js
@@ -3,9 +3,9 @@ const {
     findGroupByID,
     findGroupContainingEntryID,
     findGroupContainingGroupID
-} = require("./tools/rawVaultSearch.js");
-const { encodeStringValue } = require("./tools/encoding.js");
-const { generateEntryHistoryItem } = require("./tools/history.js");
+} = require("../../tools/rawVaultSearch.js");
+const { encodeStringValue } = require("../../tools/encoding.js");
+const { generateEntryHistoryItem } = require("../../tools/history.js");
 
 function executeArchiveID(archive, opts, id) {
     if (opts.shareID) {
